+++ src/components/DrinkingSessionOverview/types.ts
@@ -1,0 +1,9 @@
+import type {DrinkingSession, DrinkingSessionId} from '@src/types/onyx';
+
+type DrinkingSessionOverviewProps = {
+  sessionId: DrinkingSessionId;
+  session: DrinkingSession;
+  isEditModeOn: boolean;
+};
+
+export type {DrinkingSessionOverviewProps};
