+++ src/libs/migrateOnyx.ts
@@ -12,14 +12,14 @@ export default function () {
 
   return new Promise<void>(resolve => {
     // Add all migrations to an array so they are executed in order
-    const migrationPromises = [
-      //   CheckForPreviousReportActionID,
-      //   RenameReceiptFilename,
-      //   KeyReportActionsDraftByReportActionID,
-      //   TransactionBackupsToCollection,
-      //   RemoveEmptyReportActionsDrafts,
-      //   NVPMigration,
-    ];
+    // const migrationPromises = [
+    //   CheckForPreviousReportActionID,
+    //   RenameReceiptFilename,
+    //   KeyReportActionsDraftByReportActionID,
+    //   TransactionBackupsToCollection,
+    //   RemoveEmptyReportActionsDrafts,
+    //   NVPMigration,
+    // ];
 
     // Reduce all promises down to a single promise. All promises run in a linear fashion, waiting for the
     // previous promise to finish before moving onto the next one.
