+++ src/styles/utils/spacing.ts
@@ -11,6 +11,10 @@ export default {
     margin: 0,
   },
 
+  m1: {
+    margin: 4,
+  },
+
   m2: {
     margin: 8,
   },
